+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Un-Released]
 
+## [2.2.7]
+- New features:
+    - Add advance in cancel tx
+    - Add ZkApp multiple transactions at the same time
+    - Add ZkApp switch chain
+    - Add ZkApp add chain
+    - Add ZkApp sign json messgae
+    - Add nullifier support
+    - Add zk-commond parse
+- Network config enhancements 
+
+
 ## [2.2.6]
 - Add support for Testworld2.0 network